If file is found but cannot be saved, delete it from the database

+++ ingest/plugins.py
@@ -2295,7 +2295,8 @@ class CovidPolicyPlugin(IngestPlugin):
                     else:
                         print('Could not download file at URL ' +
                               str(file_url))
-                        file.delete()
+                        if file is not None:
+                            file.delete()
                         commit()
                         could_not_download.add(file_url)
                         n_failed += 1
